Refactor error handling for mesh texture loading
A try-except block has been added for error handling around the texture loading for 3D meshes in the d_graphml module. This is to prevent the code from breaking when the texture file is missing or can't be read. If the texture can't be found, the mesh will be added to the plotter without any texture.