Adjust code for logo handling and model reading
The splash and logo_widget calls have been commented out in EDMatrix2Graphml.py and load_3d.py respectively. Additional error handling has been put in place in load_3d.py where it will print a failure message if it fails to read the model from the obj_path. Also, the intensity of the light in the plotter was increased from 0.7 to 1.