Add click callback functionality to load_3d module
The update introduces a click event handler in the load_3d module. This function is triggered upon a click event on an object within the 3D representation, and it retrieves the clicked object's position to select a cell in the 3D renderer. Additionally, a minor change has been made on a comment to make it more comprehensive and clear.