Implement update status labels and table refresh functionalities
A substantial update has been made on EDMatrix2Graphml.py by adding a method to update status labels, declaring the current record index, count of total records, and sort status. Other methods were also implemented to return the current record index and total record count, as well as to refresh the data table when needed. The removal or update of data within the data table was also incorporated on a row-specific basis which is a significant improvement in the code base.