Refactor listWidget name in EDMatrix2Graphml.py
The main action in this commit is the renaming of the listWidget_list_media object to custom_list_widget in the EDMatrix2Graphml.py file. This renaming is reflected in two lines where the object is used, including where it's cleared and where items are added to it. Additionally, redundant error logs in log.txt have been removed.